# perf: optimize performance