--- a/__tests__/demos/perf/index.ts
+++ b/__tests__/demos/perf/index.ts
@@ -1,1 +1,2 @@
 export { circles } from './circles';
+export { rects } from './rect';