--- a/__tests__/demos/perf/index.ts
+++ b/__tests__/demos/perf/index.ts
@@ -1,2 +1,3 @@
 export { circles } from './circles';
 export { rects } from './rect';
+export { customElement } from './custom';