--- a/packages/g-lite/src/AbstractRenderer.ts
+++ b/packages/g-lite/src/AbstractRenderer.ts
@@ -85,7 +85,6 @@ export class AbstractRenderer implements IRenderer {
        * enable dirty rectangle rendering by default
        */
       enableDirtyRectangleRendering: true,
-      enableDirtyRectangleRenderingDebug: false,
       enableSizeAttenuation: true,
       imageSmoothingEnabled: true,
       imageSmoothingQuality: 'high',