--- a/packages/g-lite/src/css/CSSProperty.ts
+++ b/packages/g-lite/src/css/CSSProperty.ts
@@ -71,5 +71,5 @@ export interface CSSProperty<Parsed, Used> {
   /**
    * eg. update local position after x/y/z caculated
    */
-  postProcessor: (object: IElement, attributes: string[]) => void;
+  postProcessor: (object: IElement) => void;
 }