--- a/packages/g-lite/src/css/parser/transform.ts
+++ b/packages/g-lite/src/css/parser/transform.ts
@@ -137,6 +137,7 @@ export function parseTransform(
   const result: ParsedTransform[] = [];
   let match;
   let prevLastIndex = 0;
+  TRANSFORM_REGEXP.lastIndex = 0;
   while ((match = TRANSFORM_REGEXP.exec(transform))) {
     if (match.index !== prevLastIndex) {
       return [];