--- a/packages/g-lite/src/index.ts
+++ b/packages/g-lite/src/index.ts
@@ -1,3 +1,5 @@
+import './preset';
+
 // @see https://github.com/antvis/G/issues/1239
 import RBush from 'rbush/rbush.js';
 export * from './AbstractRenderer';