… .changeset/cool-crews-sneeze.md
… .github/workflows/test.yml
… __tests__/index.html
… __tests__/integration/snapshots/2d/canvas/circle.png
… __tests__/integration/snapshots/2d/canvas/clipPath.png
… __tests__/integration/snapshots/2d/canvas/customElement.png
… __tests__/integration/snapshots/2d/canvas/ellipse.png
… __tests__/integration/snapshots/2d/canvas/getPoint.png
… __tests__/integration/snapshots/2d/canvas/gradient.png
… __tests__/integration/snapshots/2d/canvas/html.png
… __tests__/integration/snapshots/2d/canvas/image.png
… __tests__/integration/snapshots/2d/canvas/line.png
… __tests__/integration/snapshots/2d/canvas/marker.png
… __tests__/integration/snapshots/2d/canvas/path.png
… __tests__/integration/snapshots/2d/canvas/pattern.png
… __tests__/integration/snapshots/2d/canvas/pattern2.png
… __tests__/integration/snapshots/2d/canvas/polygon.png
… __tests__/integration/snapshots/2d/canvas/polyline.png
… __tests__/integration/snapshots/2d/canvas/rect.png
… __tests__/integration/snapshots/2d/canvas/text.png
… __tests__/integration/snapshots/2d/canvas/transform.png
… __tests__/integration/snapshots/2d/canvas/transformOrigin.png
… __tests__/integration/snapshots/2d/canvas/transformSkew.png
… __tests__/integration/snapshots/2d/canvas/transformText.png
… __tests__/integration/snapshots/2d/canvas/zIndex.png
… __tests__/integration/snapshots/2d/svg/circle.png
… __tests__/integration/snapshots/2d/svg/clipPath.png
… __tests__/integration/snapshots/2d/svg/customElement.png
… __tests__/integration/snapshots/2d/svg/ellipse.png
… __tests__/integration/snapshots/2d/svg/getPoint.png