-  [x] 设置 glMatrix Array Type 为 Array

实测 10W Group+ 10W Rect 下，绘制时长稳定提升约 `2.7%` (1120ms -> 1090ms)

- [x] 优化 vec3.equals 性能

性能提升约 5%

<img width="981" alt="image" src="https://github.com/user-attachments/assets/9d3b46af-2d32-44b1-ab37-5c5c480ed5bc">

- [x] Transform 支持传入数组类型

实测 10W Group+ 10W Rect 下，绘制时长稳定提升约 `12%` (1070ms -> 940ms)

- [x] ~~支持批量插入节点~~

使用方式：

```ts
import { DocumentFragment, Group, Circle } from '@antv/g';

const group = new Group();
const fragment = new DocumentFragment();

for (let i = 0; i < 10_000; i++) {
  const circle = new Circle({ style: { cx: 10, cy: 10, r: 10 } });
  fragment.appendChild(circle);
}

// 批量插入
group.appendChild(fragment);
```

实测 3层 Group， 10W Group+ 10W Rect 下，绘制时长稳定提升约 `12%` (1180ms -> 1030ms)

- [x] 推迟首屏事件触发及 RTree 同步

优化首屏 triggerPendingEvents ，首屏下一帧进行全量同步，并跳过首屏帧内的事件更新

实测 3层 Group， 10W Group+ 10W Rect 下，绘制时长稳定提升约 `18%` (1050ms -> 860ms)

> 关闭性能工具的情况下，绘制时长甚至可以控制在 `600ms` 内

- [x] 优化循环调用
- [x] 提供 Canvas 抗锯齿开关
- [x] 优化代码逻辑，减少 GC 耗时 

实测 3层 Group， 10W Group+ 10W Rect 下，绘制时长稳定提升约 `30%` (860ms -> 610ms)

- 新增 `fastCleanExistingCanvas` 选项，销毁画布过程中跳过元素生命周期操作，提高销毁速度

⚠️ 已知问题：
* WebGL  polygon 案例闪烁无法绘制，原因未知
* WebGL pattern2 抛出异常