@Aarebecca 2024-10-22
alternative: https://github.com/antvis/G/pull/1789